single type

+++ python_bladerf/pybladerf_tools/pybladerf_sweep.pyx
@@ -145,13 +145,13 @@ cdef void process_data(object device):
                         'timestamp': time_str,
                         'start_frequency': frequency,
                         'stop_frequency': frequency + sample_rate // 4,
-                        'array': pwr[fft_1_start:fft_1_stop]
+                        'array': pwr[fft_1_start:fft_1_stop].astype(np.float32)
                     })
                     current_device_data['queue'].put({
                         'timestamp': time_str,
                         'start_frequency': frequency + sample_rate // 2,
                         'stop_frequency': frequency + (sample_rate * 3) // 4,
-                        'array': pwr[fft_2_start:fft_2_stop]
+                        'array': pwr[fft_2_start:fft_2_stop].astype(np.float32)
                     })
 
                 else:
@@ -159,7 +159,7 @@ cdef void process_data(object device):
                         'timestamp': time_str,
                         'start_frequency': frequency,
                         'stop_frequency': frequency + sample_rate,
-                        'array': pwr
+                        'array': pwr.astype(np.float32)
                     })
 
             else:
